Adding my test program
Couldnt upload tensorflow api it's too big. right now im anle to input data and get output. Im also able to test how good the output is and train it. not included in the test file I have also found a way to save the weights this way if we train it for a while we can checkpoint it and observe how the net outputs over time